Data.VMap

Documentation

class MVector (Mutable v) a => Vector (v :: Type -> Type) a

Class of immutable vectors. Every immutable vector is associated with its mutable version through the Mutable type family. Methods of this class should not be used directly. Instead, Data.Vector.Generic and other Data.Vector modules provide safe and fusible wrappers.

Minimum complete implementation:

basicUnsafeFreeze

basicUnsafeThaw

basicLength

basicUnsafeSlice

basicUnsafeIndexM

Minimal complete definition

basicUnsafeFreeze, basicUnsafeThaw, basicLength, basicUnsafeSlice, basicUnsafeIndexM

type VB = Vector

type VU = Vector

type VP = Vector

type VS = Vector

newtype VMap kv vv k v

Constructors

VMap

Fields

unVMap :: KVVector kv vv (k, v)

empty :: (Vector kv k, Vector vv v) => VMap kv vv k v

size :: Vector kv k => VMap kv vv k v -> Int

lookup :: (Ord k, Vector kv k, Vector vv v) => k -> VMap kv vv k v -> Maybe v

findWithDefault :: (Ord k, Vector kv k, Vector vv v) => v -> k -> VMap kv vv k v -> v

member :: (Ord k, Vector kv k) => k -> VMap kv vv k v -> Bool

notMember :: (Ord k, Vector kv k) => k -> VMap kv vv k v -> Bool

map :: (Vector vv a, Vector vv b) => (a -> b) -> VMap kv vv k a -> VMap kv vv k b

mapMaybe :: (Vector kv k, Vector vv a, Vector vv b) => (a -> Maybe b) -> VMap kv vv k a -> VMap kv vv k b

mapWithKey :: (Vector kv k, Vector vv a, Vector vv b) => (k -> a -> b) -> VMap kv vv k a -> VMap kv vv k b

filter :: (Vector kv k, Vector vv v) => (k -> v -> Bool) -> VMap kv vv k v -> VMap kv vv k v

fold :: (Vector vv m, Monoid m) => VMap kv vv k m -> m

Fold values monoidally

foldl :: Vector vv v => (a -> v -> a) -> a -> VMap kv vv k v -> a

foldlWithKey :: (Vector kv k, Vector vv v) => (a -> k -> v -> a) -> a -> VMap kv vv k v -> a

foldMap :: (Vector vv v, Monoid m) => (v -> m) -> VMap kv vv k v -> m

foldMapWithKey :: (Vector kv k, Vector vv v, Monoid m) => (k -> v -> m) -> VMap kv vv k v -> m

fromMap :: (Vector kv k, Vector vv v) => Map k v -> VMap kv vv k v

toMap :: (Vector kv k, Vector vv v) => VMap kv vv k v -> Map k v

fromList :: (Ord k, Vector kv k, Vector vv v) => [(k, v)] -> VMap kv vv k v

fromListN :: (Ord k, Vector kv k, Vector vv v) => Int -> [(k, v)] -> VMap kv vv k v

toList :: (Vector kv k, Vector vv v) => VMap kv vv k v -> [(k, v)]

toAscList :: (Vector kv k, Vector vv v) => VMap kv vv k v -> [(k, v)]

keys :: Vector kv k => VMap kv vv k v -> [k]

elems :: Vector vv v => VMap kv vv k v -> [v]

fromAscList :: (Eq k, Vector kv k, Vector vv v) => [(k, v)] -> VMap kv vv k v

fromAscListN :: (Eq k, Vector kv k, Vector vv v) => Int -> [(k, v)] -> VMap kv vv k v

fromAscListWithKey :: (Eq k, Vector kv k, Vector vv v) => (k -> v -> v -> v) -> [(k, v)] -> VMap kv vv k v

fromAscListWithKeyN :: (Eq k, Vector kv k, Vector vv v) => Int -> (k -> v -> v -> v) -> [(k, v)] -> VMap kv vv k v

fromDistinctAscList :: (Vector kv k, Vector vv v) => [(k, v)] -> VMap kv vv k v

fromDistinctAscListN :: (Vector kv k, Vector vv v) => Int -> [(k, v)] -> VMap kv vv k v

internMaybe :: (Vector kv k, Ord k) => k -> VMap kv vv k v -> Maybe k

null :: (Vector vv v, Vector kv k) => VMap kv vv k v -> Bool

splitAt :: (Vector vv v, Vector kv k) => Int -> VMap kv vv k v -> (VMap kv vv k v, VMap kv vv k v)

data KVMVector kmv vmv s a

data KVVector kv vv a

normalize :: (Vector kv k, Vector vv v, Ord k) => KVVector kv vv (k, v) -> KVVector kv vv (k, v)

normalizeM :: (Ord k, PrimMonad m, Vector kv k, Vector vv v) => KVMVector (Mutable kv) (Mutable vv) (PrimState m) (k, v) -> m (KVMVector (Mutable kv) (Mutable vv) (PrimState m) (k, v))