Data.VMap.KVVector

Documentation

class MVector (Mutable v) a ⇒ Vector (v ∷ Type → Type) a

Minimal complete definition

basicUnsafeFreeze, basicUnsafeThaw, basicLength, basicUnsafeSlice, basicUnsafeIndexM

class MVector (v ∷ Type → TYPE LiftedRep → TYPE LiftedRep) a

Minimal complete definition

basicLength, basicUnsafeSlice, basicOverlaps, basicUnsafeNew, basicInitialize, basicUnsafeRead, basicUnsafeWrite

data KVVector kv vv a

Constructors

KVVector

Fields

keysVector ∷ !(kv (Key a))

valsVector ∷ !(vv (Value a))

data KVMVector kmv vmv s a

toMap ∷ (Vector kv k, Vector vv v) ⇒ KVVector kv vv (k, v) → Map k v

Convert a sorted key/value vector into a Map

fromMap ∷ (Vector kv k, Vector vv v) ⇒ Map k v → KVVector kv vv (k, v)

Convert a Map into a sorted key/value vector.

fromAscList ∷ (Eq k, Vector kv k, Vector vv v) ⇒ [(k, v)] → KVVector kv vv (k, v)

Convert a sorted assoc list into a KVVector

fromAscListN ∷ (Eq k, Vector kv k, Vector vv v) ⇒ Int → [(k, v)] → KVVector kv vv (k, v)

Convert a sorted assoc list into a KVVector

fromAscListWithKey ∷ (Eq k, Vector kv k, Vector vv v) ⇒ (k → v → v → v) → [(k, v)] → KVVector kv vv (k, v)

fromAscListWithKeyN ∷ (Eq k, Vector kv k, Vector vv v) ⇒ Int → (k → v → v → v) → [(k, v)] → KVVector kv vv (k, v)

fromDistinctAscList ∷ (Vector kv k, Vector vv v) ⇒ [(k, v)] → KVVector kv vv (k, v)

Convert a sorted assoc list with distionct keys into a KVVector

fromDistinctAscListN ∷ (Vector kv k, Vector vv v) ⇒ Int → [(k, v)] → KVVector kv vv (k, v)

Convert a sorted assoc list with distionct keys into a KVVector. Length must be supplied.

fromList ∷ (Vector kv k, Vector vv v, Ord k) ⇒ [(k, v)] → KVVector kv vv (k, v)

Convert a possibly unsorted assoc list into a KVVector.

fromListN ∷ (Vector kv k, Vector vv v, Ord k) ⇒ Int → [(k, v)] → KVVector kv vv (k, v)

Convert a possibly unsorted assoc list into a KVVector.

mapValsKVVector ∷ (Vector vv a, Vector vv b) ⇒ (a → b) → KVVector kv vv (k, a) → KVVector kv vv (k, b)

mapWithKeyKVVector ∷ (Vector kv k, Vector vv a, Vector vv b) ⇒ (k → a → b) → KVVector kv vv (k, a) → KVVector kv vv (k, b)

memberKVVector ∷ (Ord k, Vector kv k) ⇒ k → KVVector kv vv (k, v) → Bool

lookupKVVector ∷ (Ord k, Vector kv k, Vector vv v) ⇒ k → KVVector kv vv (k, v) → Maybe v

Look up a value by the key in a sorted key/value vector. Ensure it is sorted otherwise terrible things happen.

lookupDefaultKVVector ∷ (Ord k, Vector kv k, Vector vv v) ⇒ v → k → KVVector kv vv (k, v) → v

Look up a value by the key in a sorted key/value vector. Ensure it is sorted otherwise terrible things happen.

sortAscKVMVector ∷ (MVector kmv k, MVector vmv v, Ord k, PrimMonad m) ⇒ KVMVector kmv vmv (PrimState m) (k, v) → m ()

internKVVectorMaybe ∷ (Vector kv k, Ord k) ⇒ k → KVVector kv vv (k, v) → Maybe k

normalize ∷ (Vector kv k, Vector vv v, Ord k) ⇒ KVVector kv vv (k, v) → KVVector kv vv (k, v)

normalizeM ∷ (Ord k, PrimMonad m, Vector kv k, Vector vv v) ⇒ KVMVector (Mutable kv) (Mutable vv) (PrimState m) (k, v) → m (KVMVector (Mutable kv) (Mutable vv) (PrimState m) (k, v))

Orphan instances

Typeable e ⇒ NoThunks (Vector e)

Instance details

Methods

noThunks ∷ Context → Vector e → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → Vector e → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (Vector e) → String

Typeable e ⇒ NoThunks (Vector e)

Instance details

Methods

noThunks ∷ Context → Vector e → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → Vector e → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (Vector e) → String